requirement.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
oauth2client
ezgmail
pysocks
flask
gevent
requests[socks]
toml
chardet